# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Live Score — Housing Trust Infrastructure API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leases:
    get:
      operationId: listLeases
      tags: [leases]
      summary: List current user's leases
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Lease list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lease"
    post:
      operationId: createLease
      tags: [leases]
      summary: Add a new lease
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaseInput"
      responses:
        "201":
          description: Created lease
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lease"
        "401":
          description: Authentication required
        "422":
          description: Lease validation failed
  /leases/{id}:
    patch:
      operationId: updateLease
      tags: [leases]
      summary: Update lease — landlord rating, deposit returned, end date
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaseUpdate"
      responses:
        "200":
          description: Updated lease
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lease"
        "400":
          description: Invalid lease input
        "401":
          description: Authentication required
        "404":
          description: Lease not found
        "422":
          description: Lease validation failed
  /leases/{id}/confirm-payment:
    post:
      operationId: confirmPayment
      tags: [leases]
      summary: Record an on-time payment — awards +15 score (streak bonus every 6)
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "401":
          description: Authentication required
        "404":
          description: Lease not found
  /leases/{id}/late-payment:
    post:
      operationId: recordLatePayment
      tags: [leases]
      summary: Record a late payment — penalises -25 score
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Late payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "401":
          description: Authentication required
        "404":
          description: Lease not found
components:
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: false
      schema:
        type: string
        minLength: 1
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Lease:
      type: object
      required: [id, address, city, monthlyRentRub, startDate, isActive, onTimePayments, latePayments]
      properties:
        id:
          type: integer
        address:
          type: string
        city:
          type: string
        landlordName:
          type: ["string", "null"]
        landlordUserId:
          type: ["integer", "null"]
        monthlyRentRub:
          type: number
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        isActive:
          type: boolean
        depositAmountRub:
          type: ["number", "null"]
        depositReturned:
          type: ["boolean", "null"]
        onTimePayments:
          type: integer
        latePayments:
          type: integer
        landlordRating:
          type: ["number", "null"]
        createdAt:
          type: string
...
    LeaseInput:
      type: object
      required: [address, monthlyRentRub]
      properties:
        address:
          type: string
        city:
          type: string
        landlordName:
          type: string
        landlordUserId:
          type: integer
        monthlyRentRub:
          type: number
          minimum: 0
        startDate:
          type: string
          format: date-time
        depositAmountRub:
          type: number
          minimum: 0
...
    PaymentResponse:
      type: object
      required: [success, scoreChange, newScore, idempotent]
      properties:
        success:
          type: boolean
        scoreChange:
          type: integer
        newScore:
          type: integer
        streakBonus:
          type: integer
        onTimePayments:
          type: integer
        latePayments:
          type: integer
        idempotent:
          type: boolean
    LeaseUpdate:
      type: object
      properties:
        landlordRating:
          type: number
          minimum: 1
          maximum: 5
        landlordName:
          type: string
        depositReturned:
          type: boolean
        isActive:
          type: boolean
        endDate:
          type: string
          format: date-time
...